moved filename validation methods to a new class, created skeleton factory for file readers
