+++ src/main/java/com/sparta/rbf/employees_project2/jdbc/file_input/FileReading.java
@@ -1,0 +1,5 @@
+package com.sparta.rbf.employees_project2.jdbc.file_input;
+
+public interface FileReading {
+    void getEmployeesFromFile();
+}
