+++ src/main/java/com/sparta/rbf/employees_project2/jdbc/file_input/ReadingJSON.java
@@ -1,0 +1,8 @@
+package com.sparta.rbf.employees_project2.jdbc.file_input;
+
+public class ReadingJSON implements FileReading {
+    @Override
+    public void getEmployeesFromFile() {
+
+    }
+}
