+++ src/main/java/com/sparta/rbf/employees_project2/jdbc/file_input/ReadingXML.java
@@ -1,0 +1,8 @@
+package com.sparta.rbf.employees_project2.jdbc.file_input;
+
+public class ReadingXML implements FileReading{
+    @Override
+    public void getEmployeesFromFile() {
+
+    }
+}
