+++ src/main/java/com/sparta/rbf/employees_project2/jdbc/jackson/FileFactory.java
@@ -1,8 +1,15 @@
 package com.sparta.rbf.employees_project2.jdbc.jackson;
 
+import java.util.logging.Logger;
+
 import static com.sparta.rbf.employees_project2.jdbc.App.getFileName;
+import static com.sparta.rbf.employees_project2.jdbc.jackson.FileValidation.*;
 
 public class FileFactory {
+    /*
+    create a new class called FileValidation, move all filename validator methods to this new class
+     */
+    public static final Logger logger = Logger.getLogger(FileFactory.class.getName());
     public static void createFile(Employees employees) {
         String fileName = getFileName();
         String fileExtension = getFileExtension(fileName).toLowerCase();
@@ -11,27 +18,6 @@ public class FileFactory {
         } else {
             System.out.println("File name is invalid");
         }
-    }
-    public static boolean isFileNameValid(String filename) {
-        String regex_pattern = "^[A-za-z0-9.-]{1,255}$";
-        boolean hasValidChars = filename.matches(regex_pattern);
-        int count = 0;
-
-        for (char c : filename.toCharArray()) {
-            if (c == '.') {
-                count++;
-            }
-        }
-        return hasValidChars && count == 1;
-    }
-
-    public static boolean isFileExtensionValid(String fileExtension) {
-        return fileExtension.equals("xml") || fileExtension.equals("json");
-    }
-
-    public static String getFileExtension(String fileName) {
-        String[] fileNameExtension = fileName.split("\\.");
-        return fileNameExtension[fileNameExtension.length - 1];
     }
 
     public static void getMapper(String fileExtension, String fileName, Employees employees) {
